feat: Create database with models
A database to store data to display on a website. Using the Django
built-in Object Relational Mapper (ORM) to create classes that
correspond to database tables. ORM  Class attributes correspond to
columns, and instances of the classes correspond to rows in the
database. In ORM, the classes represent database tables and are referred
to as models. In Django, models are in models.py. In the projects app,
there is one table to store the different projects to display to the
user. Therefore only one model in models.py.
The model created is called Project and has the following fields:
- title, a short string field to hold the name of project.
- description, a larger string field to hold a longer piece of text.
- technology, a string field limited to a select number of choices.
- image, an image field holds the file path where the image is stored.
To start the process of creating  database, migration necessary.
A migration is a file containing a Migration class with rules that tell
Django what changes need to be made to the database.
A file called db.sqlite3 created in the root of project. To create
instances of Project class, use the Django shell. The Django shell is
similar to the Python shell but allows you to access the database and
create entries.